% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRHO2_initialize.R
\name{RRHO2_initialize}
\alias{RRHO2_initialize}
\title{RRHO2}
\usage{
RRHO2_initialize(list1, list2, stepsize = defaultStepSize(list1, list2),
  labels = NULL, log10.ind = FALSE, multipleTesting = "none",
  boundary = 0.1, method = "hyper")
}
\arguments{
\item{list1}{data.frame. First column is the element (possibly gene) identifier, and the second is its value on which to sort. For differential gene expression, values are often -log10(P-value) * sign(effect).}

\item{list2}{data.frame. Same as list1.}

\item{stepsize}{Controls the resolution of the test: how many items between any two overlap tests.}

\item{labels}{A two element vector indicating the label of list1 and list2.}

\item{log10.ind}{Logical. Should pvalues be reported and plotted in -log10 scale and not -log scale?}

\item{multipleTesting}{Three options. none: raw p-value; BH: BH correction; BY: BY correction.}

\item{boundary}{Size of the white strip. 0.1 indicates 10\% of the heatmap size.}

\item{method}{Method for odds ratio or pvalue representation "fisher" used odds ratio and "hyper" uses p-value.}
}
\value{
list of result
\item{hypermat}{Matrix of -log(pvals) of the test for the first i,j elements of the lists##' the overlapping test between the first ith element of the sorted list1, 
using the Stratified representation.}
\item{genelist_uu}{Genes are are up-regulated in list 1 and up-regulated list 2, at the most significant pixel of the uu quadrant}
\item{genelist_dd}{Genes are are down-regulated in list 1 and down-regulated list 2, at the most significant pixel of the dd quadrant}
\item{genelist_du}{Genes are are down-regulated in list 1 and up-regulated list 2, at the most significant pixel of the du quadrant}
\item{genelist_ud}{Genes are are up-regulated in list 1 and down-regulated list 2, at the most significant pixel of the ud quadrant}
}
\description{
An improved version for RRHO, which aims to correct the intepretation for top left region (up in x and down in y) nad bottom right region.
}
\details{
We improved the algorithm such that all four regions of RRHO plot are meaningful
}
\examples{

## A total of 2000 genes in both list 1 and list 2.
## In list 1, genes 1-200 are up-regulated; genes 201-400 are down-regulated; the rest of the 1600 genes are noise genes. 
## In list 2, genes 1-200 are up-regulated; genes 201-400 are down-regulated; the rest of the 1600 genes are noise genes.  

set.seed(15213)
nGenes <- 2000
nDE <- 200
Genes <- paste0("Genes",1:nGenes)

list1_pvalue_1_200 <- runif(nDE,0,0.05)
list1_pvalue_201_400 <- runif(nDE,0,0.05) 
list1_pvalue_401_2000 <- runif(nGenes - 2 * nDE,0,1)
list1_DDE <- c(-log10(list1_pvalue_1_200), -log10(list1_pvalue_201_400) * (-1), -log10(list1_pvalue_401_2000) * sample(c(1,-1), length(list1_pvalue_401_2000), replace = TRUE))

gene_list1 <- data.frame(Genes=Genes,DDE = list1_DDE, stringsAsFactors = FALSE)

list2_pvalue_1_200 <- runif(nDE,0,0.05)
list2_pvalue_201_400 <- runif(nDE,0,0.05) 
list2_pvalue_401_2000 <- runif(nGenes - 2 * nDE,0,1)
list2_DDE <- c(-log10(list2_pvalue_1_200), -log10(list2_pvalue_201_400) * (-1), -log10(list2_pvalue_401_2000) * sample(c(1,-1), length(list2_pvalue_401_2000), replace = TRUE))

gene_list2 <- data.frame(Genes=Genes,DDE = list2_DDE, stringsAsFactors = FALSE)

RRHO_obj <-  RRHO2_initialize(gene_list1, gene_list2, labels = c("list1", "list2"), log10.ind=TRUE)

}
\author{
Kelly and Caleb
}
