% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRHO2.R
\name{RRHO2}
\alias{RRHO2}
\title{RRHO2}
\usage{
RRHO2(list1, list2, stepsize = defaultStepSize(list1, list2), labels,
  plots = FALSE, outputdir = NULL, BY = FALSE, log10.ind = FALSE,
  maximum = 50, boundary = 0.1, res = 100, method = "fisher",
  alternative = "split")
}
\arguments{
\item{RRHO_obj}{RRHO object. See RRHO2_initialize for details.}
}
\description{
An improved version for RRHO, which aims to correct the intepretation for top left region (up in x and down in y) nad bottom right region.
}
\details{
We improved the algorithm such that all four regions of RRHO plot are meaningful
}
\examples{

plotFolder <- 'plot'
system(paste('mkdir -p', plotFolder))
list.length <- 2000
list.names <- paste('Gene',1:list.length, sep='')
set.seed(15213)
gene.list1<- data.frame(Genes=list.names, Value = sample(list.length)*sample(c(1,-1),list.length,replace=TRUE))
gene.list2<- data.frame(Genes=list.names, Value = sample(list.length)*sample(c(1,-1),list.length,replace=TRUE))
# Enrichment alternative
RRHO.example <-  RRHO2(gene.list1, gene.list2, 
                      labels=c('x','y'), plots=TRUE, outputdir=plotFolder, BY=TRUE, log10.ind=TRUE)

}
\author{
Kelly and Caleb
}
