% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRHO2.R
\name{RRHO2}
\alias{RRHO2}
\title{RRHO2}
\usage{
RRHO2(list1, list2, stepsize = defaultStepSize(list1, list2), labels,
  plots = FALSE, outputdir = NULL, BY = FALSE, log10.ind = FALSE)
}
\arguments{
\item{list1}{data.frame. First column is the element (possibly gene) identifier, and the second is its value on which to sort. For differential gene expression, values are often -log10(P-value) * sign(effect).}

\item{list2}{data.frame. Same as list1.}

\item{stepsize}{Controls the resolution of the test: how many items between any two overlap tests.}

\item{labels}{Character vector with two elements: the labels of the two lists.}

\item{plots}{Logical. Should output plots be returned?}

\item{outputdir}{Path name where plots ae returned.}

\item{BY}{Logical. Should Benjamini-Yekutieli FDR corrected pvalues be computed?}

\item{log10.ind}{Logical. Should pvalues be reported and plotted in -log10 scale and not -log scale?}
\item{size}{Controls the height and width of the square dimensions of the heatmap. Default is 8.} 
\item{fill}{Vector. Vector of two colors for venn diagrams. Default is c("cornflowerblue", "darkorchid1").}
}
\value{
list of result
\item{hypermat}{Matrix of -log(pvals) of the test for the first i,j elements of the lists.}
}
\description{
An improved version for RRHO, which aims to correct the intepretation for top left region (up in x and down in y) nad bottom right region.
}
\details{
We improved the algorithm such that all four regions of RRHO plot are meaningful
}
\examples{

plotFolder <- 'plot'
system(paste('mkdir -p', plotFolder))
list.length <- 100
list.names <- paste('Gene',1:list.length, sep='')
set.seed(15213)
gene.list1<- data.frame(list.names, sample(100)*sample(c(1,-1),100,replace=TRUE))
gene.list2<- data.frame(list.names, sample(100)*sample(c(1,-1),100,replace=TRUE))
# Enrichment alternative
RRHO.example <-  RRHO2(gene.list1, gene.list2, 
                      labels=c('x','y'), plots=TRUE, outputdir=plotFolder, BY=TRUE, log10.ind=TRUE)

}
\author{
Caleb
}

